requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
python-dotenv
google-generativeai
reportlab

sqlalchemy
psycopg2-binary